Added to and updated the unit tests
